# backend/src/main/java/com/tribly/api/comments/RideCommentResource.java
package com.tribly.api.comments;

import com.tribly.domain.common.TeamEntity;
import com.tribly.domain.team.Team;
import com.tribly.domain.user.User;
import com.tribly.dto.comments.request.CommentRequest;
import com.tribly.dto.comments.response.CommentDto;
import com.tribly.dto.comments.response.CommentListResponse;
import com.tribly.dto.error.ErrorResponse;
import com.tribly.enums.ActionType;
import com.tribly.service.ride.RideService;
import jakarta.inject.Inject;
import jakarta.ws.rs.DELETE;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.core.Response;
import org.eclipse.microprofile.openapi.annotations.Operation;
import org.eclipse.microprofile.openapi.annotations.media.Content;
import org.eclipse.microprofile.openapi.annotations.media.Schema;
import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
import org.eclipse.microprofile.openapi.annotations.tags.Tag;

@Path("/api/teams/{slug}/rides/{entitySlug}/comments")
@Tag(name = "Ride Comments", description = "Comment operations for rides")
public class RideCommentResource extends AbstractCommentResource {

  @Inject RideService rideService;

  @Override
  protected TeamEntity getTeamEntity(Team team, String entitySlug, User user) {
    return rideService.get(ActionType.READ, team, entitySlug, user);
  }

  @Override
  protected String getEntityType() {
    return "rides";
  }

  @GET
  @Override
  @Operation(operationId = "listRideComments", summary = "List ride comments")
  @APIResponses({
    @APIResponse(
        responseCode = "200",
        description = "Comments retrieved",
        content = @Content(schema = @Schema(implementation = CommentListResponse.class))),
    @APIResponse(
        responseCode = "403",
        description = "Not a team member",
        content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
  })
  public Response listComments(
      @Parameter(description = "Team URL slug") @PathParam("slug") String teamSlug,
      @Parameter(description = "Ride URL slug") @PathParam("entitySlug") String entitySlug) {
    return super.listComments(teamSlug, entitySlug);
  }

  @POST
  @Override
  @Operation(operationId = "createRideComment", summary = "Create ride comment")
  @APIResponses({
    @APIResponse(
        responseCode = "201",
        description = "Comment created",
        content = @Content(schema = @Schema(implementation = CommentDto.class))),
    @APIResponse(
        responseCode = "400",
        description = "Invalid request",
        content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
  })
  public Response createComment(
      @Parameter(description = "Team URL slug") @PathParam("slug") String teamSlug,
      @Parameter(description = "Ride URL slug") @PathParam("entitySlug") String entitySlug,
      CommentRequest request) {
    return super.createComment(teamSlug, entitySlug, request);
  }

  @DELETE
  @Path("/{commentId}")
  @Operation(operationId = "deleteRideComment", summary = "Delete ride comment")
  @APIResponses({
    @APIResponse(responseCode = "204", description = "Comment deleted"),
    @APIResponse(
        responseCode = "403",
        description = "Cannot delete",
        content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
  })
  public Response deleteComment(
      @Parameter(description = "Team URL slug") @PathParam("slug") String teamSlug,
      @Parameter(description = "Ride URL slug") @PathParam("entitySlug") String entitySlug,
      @Parameter(description = "Comment ID") @PathParam("commentId") String commentId) {
    return super.deleteComment(teamSlug, commentId);
  }
}