# contracts/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "AssetsDto" : {
        "type" : "object",
        "properties" : {
          "logo" : {
            "description" : "Logo",
            "type" : "object",
            "$ref" : "#/components/schemas/AssetDto"
          },
          "images" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AssetDto"
            },
            "description" : "Images"
          },
          "videos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AssetDto"
            },
            "description" : "Videos"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AssetDto"
            },
            "description" : "Attachments"
          },
          "originalGpx" : {
            "description" : "Original GPX",
            "type" : "object",
            "$ref" : "#/components/schemas/AssetDto"
          },
          "gpx" : {
            "description" : "GPX",
            "type" : "object",
            "$ref" : "#/components/schemas/AssetDto"
          },
          "fit" : {
            "description" : "FIT",
            "type" : "object",
            "$ref" : "#/components/schemas/AssetDto"
          },
          "thumbnail" : {
            "description" : "Thumbnail",
            "type" : "object",
            "$ref" : "#/components/schemas/AssetDto"
          }
        },
        "required" : [ "images", "videos", "attachments" ]
      },
...
      "CommentDto" : {
        "description" : "Comment data",
        "type" : "object",
        "required" : [ "id", "content", "author", "createdAt", "replies" ],
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Comment ID (TSID)"
          },
          "content" : {
            "type" : "string",
            "description" : "Comment content"
          },
          "author" : {
            "description" : "Comment author",
            "type" : "object",
            "$ref" : "#/components/schemas/PublicUserDto"
          },
          "createdAt" : {
            "description" : "Creation timestamp",
            "type" : "string",
            "$ref" : "#/components/schemas/Instant"
          },
          "parentId" : {
            "type" : "string",
            "description" : "Parent comment ID (for replies)"
          },
          "replies" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommentDto"
            },
            "description" : "Replies to this comment"
          }
        }
      },
      "CommentListResponse" : {
        "description" : "List of comments response",
        "type" : "object",
        "required" : [ "items", "total" ],
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommentDto"
            },
            "description" : "List of comments (top-level only, with nested replies)"
          },
          "total" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Total count including replies"
          }
        }
      },
      "CommentRequest" : {
        "description" : "Comment creation request",
        "type" : "object",
        "required" : [ "content" ],
        "properties" : {
          "content" : {
            "type" : "string",
            "description" : "Comment content",
            "pattern" : "\\S",
            "maxLength" : 5000
          },
          "parentId" : {
            "type" : "string",
            "description" : "Parent comment ID for replies (optional)"
          }
        }
      },
...
      "ErrorCode" : {
        "type" : "string",
        "enum" : [ "BAD_REQUEST", "NOT_FOUND", "UNAUTHORIZED", "FORBIDDEN", "INTERNAL_ERROR", "UNKNOWN", "VALIDATION", "BUSINESS_RULE", "LAST_ADMIN", "ALREADY_REGISTERED", "NOT_REGISTERED", "INVALID_SLUG", "SLUG_TAKEN", "USER_NOT_SYNCED", "GROUP_FULL", "INVALID_VISIBILITY", "PUBLIC_TRIP_PRIVATE_ROUTE", "GPX_FAILURE", "FILE_REQUIRED", "TOO_MANY_TEAM_PAGES", "GPX_EMPTY", "INVALID_FORMAT", "RENTAL_PERIOD_MISSING", "STATUS_INVALID", "PUBLIC_RIDE_PRIVATE_ROUTE" ]
      },
      "ErrorDetails" : {
        "oneOf" : [ {
          "$ref" : "#/components/schemas/ErrorValidationDetails"
        }, {
          "$ref" : "#/components/schemas/NotFoundDetails"
        } ],
        "description" : "Error details",
        "discriminator" : {
          "propertyName" : "type",
          "mapping" : {
            "VALIDATION" : "#/components/schemas/ErrorValidationDetails",
            "NOT_FOUND" : "#/components/schemas/NotFoundDetails"
          }
        },
        "type" : "object",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/ErrorCode"
          }
        }
      },
      "ErrorResponse" : {
        "description" : "Error response",
        "type" : "object",
        "required" : [ "code" ],
        "properties" : {
          "code" : {
            "description" : "Error code",
            "type" : "string",
            "$ref" : "#/components/schemas/ErrorCode"
          },
          "errorDetails" : {
            "description" : "Additional details",
            "type" : "object",
            "$ref" : "#/components/schemas/ErrorDetails"
          }
        }
      },
      "ErrorValidationDetails" : {
        "type" : "object",
        "required" : [ "type", "fieldErrors" ],
        "properties" : {
          "type" : {
            "description" : "Type",
            "type" : "string",
            "$ref" : "#/components/schemas/ErrorCode"
          },
          "fieldErrors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            },
            "description" : "Field Errors"
          }
        }
      },
      "FieldError" : {
        "type" : "object",
        "required" : [ "field", "message", "rejectedValue" ],
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "Field"
          },
          "message" : {
            "type" : "string",
            "description" : "Message"
          },
          "rejectedValue" : {
            "description" : "Rejected value"
          }
        }
      },
...
      "GroupRequest" : {
        "description" : "Ride group creation request",
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "id"
          },
          "name" : {
            "type" : "string",
            "description" : "Group name",
            "examples" : [ "Fast Group" ],
            "pattern" : "\\S",
            "minLength" : 1,
            "maxLength" : 200
          },
          "time" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "averageSpeed" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Average speed in km/h",
            "examples" : [ 25 ],
            "exclusiveMinimum" : 0
          },
          "maxParticipants" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Maximum participants",
            "examples" : [ 10 ],
            "exclusiveMinimum" : 0
          },
          "routeSlug" : {
            "type" : "string",
            "description" : "Route slug for this group"
          }
        },
        "required" : [ "name" ]
      },
...
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
...
      "LocalTime" : {
        "type" : "string",
        "format" : "local-time",
        "examples" : [ "13:45:30.123456789" ],
        "externalDocs" : {
          "description" : "As defined by 'partial-time' in RFC3339",
          "url" : "https://www.rfc-editor.org/rfc/rfc3339.html#section-5.6"
        }
      },
      "MediaDto" : {
        "type" : "object",
        "required" : [ "markdown", "assets" ],
        "properties" : {
          "markdown" : {
            "type" : "string",
            "description" : "Markdown"
          },
          "assets" : {
            "description" : "Assets",
            "type" : "object",
            "$ref" : "#/components/schemas/AssetsDto"
          }
        }
      },
...
      "NotFoundDetails" : {
        "type" : "object",
        "required" : [ "type", "entityType", "searchedBy", "id" ],
        "properties" : {
          "type" : {
            "description" : "Type",
            "type" : "string",
            "$ref" : "#/components/schemas/ErrorCode"
          },
          "entityType" : {
            "description" : "Entity type",
            "type" : "string",
            "$ref" : "#/components/schemas/AllEntityType"
          },
          "searchedBy" : {
            "description" : "Search type",
            "type" : "string",
            "$ref" : "#/components/schemas/SearchedBy"
          },
          "id" : {
            "type" : "string",
            "description" : "id/slug"
          }
        }
      },
      "PlaceDetailDto" : {
        "type" : "object",
        "required" : [ "id", "name", "startPlace", "endPlace" ],
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Place ID (TSID)",
            "examples" : [ "0h4a8xzk8jv80" ]
          },
          "name" : {
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "link" : {
            "type" : "string"
          },
          "startPlace" : {
            "type" : "boolean"
          },
          "endPlace" : {
            "type" : "boolean"
          },
          "geometry" : {
            "$ref" : "#/components/schemas/GeoJsonPoint",
            "type" : "object"
          }
        }
      },
...
      "PublicUserDto" : {
        "type" : "object",
        "description" : "Public user information (limited fields)",
        "required" : [ "id", "displayName" ],
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "User ID (TSID)",
            "examples" : [ "0h4a8xzk8jv80" ]
          },
          "displayName" : {
            "type" : "string",
            "description" : "User display name"
          },
          "avatarUrl" : {
            "type" : "string",
            "description" : "User avatar URL"
          }
        }
      },
...
      "PublicationType" : {
        "type" : "string",
        "enum" : [ "RIDE", "POST", "TRIP" ]
      },
...
      "RideDto" : {
        "description" : "Ride summary data",
        "type" : "object",
        "required" : [ "type", "team", "id", "slug", "name", "media", "dateTime", "status", "visibility", "participantCount", "groupCount", "groups" ],
        "properties" : {
          "type" : {
            "description" : "Type",
            "type" : "string",
            "$ref" : "#/components/schemas/PublicationType"
          },
          "team" : {
            "description" : "Team",
            "type" : "object",
            "$ref" : "#/components/schemas/TeamPublicationDto"
          },
          "id" : {
            "type" : "string",
            "description" : "Publication ID (TSID)"
          },
          "slug" : {
            "type" : "string",
            "description" : "Publication URL slug"
          },
          "name" : {
            "type" : "string",
            "description" : "Publication name"
          },
          "media" : {
            "description" : "Publication media",
            "type" : "object",
            "$ref" : "#/components/schemas/MediaDto"
          },
          "dateTime" : {
            "description" : "Publication date/time",
            "type" : "string",
            "$ref" : "#/components/schemas/Instant"
          },
          "status" : {
            "description" : "Publication status",
            "type" : "string",
            "$ref" : "#/components/schemas/Status"
          },
          "visibility" : {
            "description" : "Visibility level",
            "type" : "string",
            "$ref" : "#/components/schemas/Visibility"
          },
          "publishAt" : {
            "description" : "Publication timestamp",
            "type" : "string",
            "$ref" : "#/components/schemas/Instant"
          },
          "createdAt" : {
            "description" : "Creation timestamp",
            "type" : "string",
            "$ref" : "#/components/schemas/Instant"
          },
          "routeSlug" : {
            "type" : "string",
            "description" : "Route slug"
          },
          "participantCount" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Number of participants"
          },
          "groupCount" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Number of groups"
          },
          "groups" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RideGroupDto"
            },
            "description" : "Ride groups"
          },
          "startPlace" : {
            "description" : "Start place",
            "type" : "object",
            "$ref" : "#/components/schemas/PlaceDetailDto"
          },
          "endPlace" : {
            "description" : "End place",
            "type" : "object",
            "$ref" : "#/components/schemas/PlaceDetailDto"
          }
        }
      },
      "RideGroupDto" : {
        "type" : "object",
        "description" : "Ride group information",
        "required" : [ "id", "name", "countParticipants", "participants", "sortOrder" ],
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Group ID (TSID)"
          },
          "name" : {
            "type" : "string",
            "description" : "Group name"
          },
          "time" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "routeSlug" : {
            "type" : "string",
            "description" : "Route slug"
          },
          "averageSpeed" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Average speed in km/h"
          },
          "maxParticipants" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Maximum participants"
          },
          "countParticipants" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Current number of participants"
          },
          "participants" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PublicUserDto"
            },
            "description" : "Participants, empty if not access"
          },
          "sortOrder" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Sort order"
          }
        }
      },
...
      "RideRequest" : {
        "description" : "Ride request",
        "type" : "object",
        "required" : [ "name", "media", "dateTime", "status", "visibility", "groups" ],
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "Ride name",
            "examples" : [ "Sunday Morning Ride" ],
            "pattern" : "\\S",
            "minLength" : 3,
            "maxLength" : 200
          },
          "media" : {
            "description" : "Ride media",
            "type" : "object",
            "$ref" : "#/components/schemas/MediaDto"
          },
          "dateTime" : {
            "description" : "Ride date/time",
            "type" : "string",
            "$ref" : "#/components/schemas/Instant"
          },
          "status" : {
            "description" : "Ride status",
            "type" : "string",
            "$ref" : "#/components/schemas/Status"
          },
          "visibility" : {
            "description" : "Visibility level",
            "type" : "string",
            "$ref" : "#/components/schemas/Visibility"
          },
          "routeSlug" : {
            "type" : "string",
            "description" : "Route slug"
          },
          "startPlaceId" : {
            "type" : "string",
            "description" : "Start place ID (TSID)"
          },
          "endPlaceId" : {
            "type" : "string",
            "description" : "End place ID (TSID)"
          },
          "publishAt" : {
            "description" : "Publication timestamp (for scheduled publishing)",
            "type" : "string",
            "$ref" : "#/components/schemas/Instant"
          },
          "groups" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GroupRequest"
            },
            "description" : "Ride groups to create"
          }
        }
      },
...
      "SearchedBy" : {
        "type" : "string",
        "enum" : [ "ID", "SLUG" ]
      },
...
      "Status" : {
        "type" : "string",
        "enum" : [ "DRAFT", "PUBLISHED", "CANCELLED" ]
      },
...
      "TeamPublicationDto" : {
        "type" : "object",
        "description" : "Team information",
        "required" : [ "id", "name", "slug", "visibility" ],
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Team ID (TSID)",
            "examples" : [ "0h4a8xzk8jv80" ]
          },
          "name" : {
            "type" : "string",
            "description" : "Team name"
          },
          "slug" : {
            "type" : "string",
            "description" : "Team URL slug"
          },
          "visibility" : {
            "description" : "Whether the team is public",
            "type" : "string",
            "$ref" : "#/components/schemas/Visibility"
          }
        }
      },
...
      "Visibility" : {
        "type" : "string",
        "enum" : [ "TEAM", "PUBLIC" ]
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "openIdConnectUrl" : "http://localhost:8180/realms/quarkus/.well-known/openid-configuration",
        "description" : "Authentication"
      }
    }
  },
...
  "paths" : {
...
    "/api/teams/{slug}/posts/{entitySlug}/comments/{commentId}" : {
      "delete" : {
        "summary" : "Delete post comment",
        "operationId" : "deletePostComment",
        "tags" : [ "Post Comments" ],
        "parameters" : [ {
          "description" : "Comment ID",
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Post URL slug",
          "name" : "entitySlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Comment deleted"
          },
          "403" : {
            "description" : "Cannot delete",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/teams/{slug}/rides/{entitySlug}/comments" : {
      "get" : {
        "summary" : "List ride comments",
        "operationId" : "listRideComments",
        "tags" : [ "Ride Comments" ],
        "parameters" : [ {
          "description" : "Ride URL slug",
          "name" : "entitySlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Comments retrieved",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommentListResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Not a team member",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create ride comment",
        "operationId" : "createRideComment",
        "tags" : [ "Ride Comments" ],
        "parameters" : [ {
          "description" : "Ride URL slug",
          "name" : "entitySlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CommentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Comment created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommentDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/teams/{slug}/rides/{entitySlug}/comments/{commentId}" : {
      "delete" : {
        "summary" : "Delete ride comment",
        "operationId" : "deleteRideComment",
        "tags" : [ "Ride Comments" ],
        "parameters" : [ {
          "description" : "Comment ID",
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Ride URL slug",
          "name" : "entitySlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Comment deleted"
          },
          "403" : {
            "description" : "Cannot delete",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/teams/{slug}/rides/{rideSlug}" : {
      "put" : {
        "summary" : "Update ride",
        "description" : "Update ride information. Requires organizer permissions.",
        "operationId" : "updateRide",
        "tags" : [ "Rides" ],
        "parameters" : [ {
          "description" : "Ride URL slug",
          "name" : "rideSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RideRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ride updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not authorized to update this ride",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team or ride not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      },
      "get" : {
        "summary" : "Get ride details",
        "description" : "Get detailed ride information including groups",
        "operationId" : "getRide",
        "tags" : [ "Rides" ],
        "parameters" : [ {
          "description" : "Ride URL slug",
          "name" : "rideSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ride retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Team or ride not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete ride",
        "description" : "Soft delete a ride. Requires organizer permissions.",
        "operationId" : "deleteRide",
        "tags" : [ "Rides" ],
        "parameters" : [ {
          "description" : "Ride URL slug",
          "name" : "rideSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Ride deleted successfully"
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not authorized to delete this ride",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team or ride not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
...
    "/api/teams/{slug}/routes/{entitySlug}/comments/{commentId}" : {
      "delete" : {
        "summary" : "Delete route comment",
        "operationId" : "deleteRouteComment",
        "tags" : [ "Route Comments" ],
        "parameters" : [ {
          "description" : "Comment ID",
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Route URL slug",
          "name" : "entitySlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Comment deleted"
          },
          "403" : {
            "description" : "Cannot delete",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/teams/{slug}/trips/{entitySlug}/comments/{commentId}" : {
      "delete" : {
        "summary" : "Delete trip comment",
        "operationId" : "deleteTripComment",
        "tags" : [ "Trip Comments" ],
        "parameters" : [ {
          "description" : "Comment ID",
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Trip URL slug",
          "name" : "entitySlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Comment deleted"
          },
          "403" : {
            "description" : "Cannot delete",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info" : {
    "title" : "Tribly API",
    "version" : "1.0.0",
    "description" : "API for Tribly Cycling Team Management Platform"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}